Move Strings to browser
