+++ src/browser/LocalizableStrings.ts
